+++ central/src/auth/mod.rs
@@ -1,0 +1,4 @@
+mod keycloak;
+pub(crate) mod config;
+mod authentik;
+mod authentik_test;
